feat(peg): background monitor with rolling-window depeg detector
F-PEG-2
Adds modules/risk_engine/peg_monitor.rs: a Tokio task spawned at startup
(gated by PEG_DEFENSE_ENABLED) that samples USDC/EURC/USYC prices every
PEG_MONITOR_TICK_SECS, feeds each rule's rolling buffer, and fires when
every sample inside window_seconds sits under threshold_price.

On fire it inserts a peg_events row, sets last_fired_at to throttle for
peg_fire_cooldown_secs, broadcasts a typed peg.alert SSE event, and dispatches
by action_kind. auto_execute is gated on Pro/Business tier — if the
subscriptions table is missing (A3 not merged) we warn-log and downgrade to
propose_rebalance. Real defensive-plan + auto-execute wiring is left as an
explicit TODO awaiting A3.

Migration is tightened from NUMERIC to DOUBLE PRECISION to match the codebase
convention and the sqlx scalar type used by the monitor (no new bigdecimal
dependency).

Nine unit tests cover the pure should_fire / cooldown / buffer-eviction logic
end-to-end with no DB.

+++ apps/api/migrations/0011_peg_rules.sql
@@ -10,7 +10,7 @@ CREATE TABLE IF NOT EXISTS peg_rules (
     -- NULL portfolio_id == apply to every portfolio the user owns.
     portfolio_id    UUID REFERENCES portfolios(id) ON DELETE CASCADE,
     asset           TEXT NOT NULL,
-    threshold_price NUMERIC NOT NULL CHECK (threshold_price > 0),
+    threshold_price DOUBLE PRECISION NOT NULL CHECK (threshold_price > 0),
     window_seconds  INTEGER NOT NULL DEFAULT 300 CHECK (window_seconds >= 0),
     action_kind     TEXT NOT NULL CHECK (action_kind IN ('alert','propose_rebalance','auto_execute')),
     target_asset    TEXT,
@@ -32,7 +32,7 @@ CREATE TABLE IF NOT EXISTS peg_events (
     id             UUID PRIMARY KEY DEFAULT gen_random_uuid(),
     rule_id        UUID NOT NULL REFERENCES peg_rules(id) ON DELETE CASCADE,
     asset          TEXT NOT NULL,
-    observed_price NUMERIC NOT NULL,
+    observed_price DOUBLE PRECISION NOT NULL,
     observed_at    TIMESTAMPTZ NOT NULL,
     action_taken   TEXT NOT NULL,
     rebalance_id   UUID REFERENCES rebalances(id) ON DELETE SET NULL,
